Fuse's attr.Size is a uint64. Cast to that type.
This test is failing because testify's assert.Equal wants the types of
both to be the same.  So the uint32 LHS will never equal the uint64 RHS,
even if they have the same value.

The other option would be to remove the cast and use assert.EqualValues
but that doesn't seem as clean: We know what type the LHS needs to be,
so doing the cast explicitly seems best.