Specify the server cert to use with the net/http/httptest server
With go 1.5, the certificate used in the net/http/httptest package changed,
which caused our tests to start failing, as they were hardcoded to check
against the cert used in 1.4 (which is in the fixtures/localhost.crt file).

To work on both 1.4 and 1.5, we supply the certificate to be used to the test
http server.  This is the same cert used by default in 1.5.  The private key
was added to fixtures/localhost.crt, since the httptest package doesn't expose
it.  A helper function was added to util_test.go that can be used to load the
certs into a TLS config, as needed for the httptest server.